+++ test/bootstrap/node.js
@@ -3,3 +3,4 @@ var chai = require('chai');
 global.expect = chai.expect;
 
 chai.use(require('chai-kerouac-middleware'));
+chai.use(require('chai-kerouac-mapper'));